teste-carga/requirements.txt
locust==2.20.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
requests==2.31.0
zeep==4.2.1
matplotlib==3.8.2
pandas==2.1.4
seaborn==0.13.0
numpy==1.26.2
